Extracting SBoD's types to a dedicated sbod-itf crate.
