+++ .gitignore
@@ -6,3 +6,4 @@ node_modules
 test-ledger
 .vscode/
 .cargo/
+.idea/
