+++ programs/sbod-itf/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "sbod-itf"
+version = "0.1.0"
+description = "Interface to read switchboard-on-demand aggregator data"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib", "lib"]
+name = "sbod_itf"
+
+[features]
+no-entrypoint = []
+cpi = ["no-entrypoint"]
+test-bpf = []
+debug = []
+
+[dependencies]
+anchor-lang = ">=0.28.0"
+solana-program = ">= 1.16.18"
+bytemuck = { version = "1.4.0", features = ["derive", "min_const_generics"] }
+rust_decimal = "1.32.0"
+static_assertions = "1.1.0"
