+++ programs/sbod-itf/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod accounts;
+
+use anchor_lang::prelude::*;
+
+declare_id!("SBondMDrcV3K4kxZR1HNVT7osZxAHVHgYXL5Ze1oMUv");
